+++ frame/staking/src/lib.rs
@@ -114,7 +114,7 @@
 //!
 //! Rewards must be claimed for each era before it gets too old by `$HISTORY_DEPTH` using the
 //! `payout_stakers` call. Any account can call `payout_stakers`, which pays the reward to the
-//! validator as well as its nominators. Only the [`Config::MaxNominatorRewardedPerValidator`]
+//! validator as well as its nominators. Only the [`Config::MaxExposurePageSize`]
 //! biggest stakers can claim their reward. This is to limit the i/o cost to mutate storage for each
 //! nominator's account.
 //!
@@ -228,7 +228,7 @@
 //! validator, proportional to the value staked behind the validator (_i.e._ dividing the
 //! [`own`](Exposure::own) or [`others`](Exposure::others) by [`total`](Exposure::total) in
 //! [`Exposure`]). Note that payouts are made in pages with each page capped at
-//! [`Config::MaxNominatorRewardedPerValidator`] nominators. The distribution of nominators across
+//! [`Config::MaxExposurePageSize`] nominators. The distribution of nominators across
 //! pages are unsorted and depends on the election result provided by [`Config::ElectionProvider`].
 //!
 //! All entities who receive a reward have the option to choose their reward destination through the